% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas_catalog_beta}
\alias{gwas_catalog_beta}
\title{Read counts and dispersion distribution for genes in HPA}
\usage{
gwas_catalog_beta
}
\description{
Read counts and dispersion distribution for genes in HPA
}
