% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HPA_rna_tissue_consensus}
\alias{HPA_rna_tissue_consensus}
\title{Read counts and dispersion distribution for genes in HPA}
\usage{
HPA_rna_tissue_consensus
}
\description{
Read counts and dispersion distribution for genes in HPA
}
