% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UKB.PheWAS.top_hits.PhenoSub.beta}
\alias{UKB.PheWAS.top_hits.PhenoSub.beta}
\title{Read counts and dispersion distribution for genes in HPA}
\usage{
UKB.PheWAS.top_hits.PhenoSub.beta
}
\description{
Read counts and dispersion distribution for genes in HPA
}
